requirements.txt
# PDF parsing
pdfplumber==0.10.3

# LLM integration
anthropic==0.28.0

# Data validation
pydantic==2.7.1

# Financial data
yfinance==0.2.40


# Excel output
openpyxl==3.1.2

# PDF output
reportlab==4.1.0

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# Utilities
python-dotenv==1.0.1